Improve benchmark
Get rid of initilize(), the new function is now more logical and straightforward.
Changed Alphametic addends and sum fields to hold Vec<char> instead of &str.
They are stored in the reversed order for a direct access to the last letter
in the word. In the result i got in average 20% benchmark improvement.